pyproject.toml
[project]
name = "yt-automation"
version = "0.1.0"
description = "Local YouTube content automation pipeline (ideas, scripts, metadata)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "jinja2>=3.1",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "rich>=13.0",
  "yt-dlp>=2024.12.0",
  "google-api-python-client>=2.120.0",
  "google-auth-oauthlib>=1.2.0",
]

[project.scripts]
yt-automation = "yt_automation.cli:main"

[tool.ruff]
line-length = 100
